requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.18.0
fsspec==2025.5.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
jiter==0.10.0
openai==1.82.0
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
